refactors all around
